+++ src/Control/Carrier/FossaApiClient/Internal/Core.hs
@@ -6,6 +6,7 @@ module Control.Carrier.FossaApiClient.Internal.Core (
   getLatestBuild,
   getOrganization,
   getProject,
+  getAnalyzedRevisions,
   getSignedUploadUrl,
   queueArchiveBuild,
   uploadAnalysis,
@@ -16,6 +17,7 @@ module Control.Carrier.FossaApiClient.Internal.Core (
 
 import App.Fossa.Config.Report (ReportOutputFormat)
 import App.Fossa.Container.Scan (ContainerScan)
+import App.Fossa.VendoredDependency (VendoredDependency (..))
 import App.Types (ProjectMetadata, ProjectRevision (..))
 import Control.Algebra (Has)
 import Control.Carrier.FossaApiClient.Internal.FossaAPIV1 qualified as API
@@ -63,6 +65,17 @@ getProject ::
 getProject revision = do
   apiOpts <- ask
   API.getProject apiOpts revision
+
+getAnalyzedRevisions ::
+  ( Has (Lift IO) sig m
+  , Has Diagnostics sig m
+  , Has (Reader ApiOpts) sig m
+  ) =>
+  NE.NonEmpty VendoredDependency ->
+  m [Text]
+getAnalyzedRevisions vdeps = do
+  apiOpts <- ask
+  API.getAnalyzedRevisions apiOpts vdeps
 
 uploadAnalysis ::
   ( Has (Lift IO) sig m